Avoid race condition with state updates
If I use a mutation-free implementation of flipCardByIx,
then when checkMatch() runs it sometimes evaluates a stale
copy of state, which causes a match not to be detected. By
passing checkMatch() as a callback to flipCardByIx, I avoid
this race condition, because checkMath() will be called immediately
after flipCardByIx() sets state.